Behave like --no-eslintrc if custom config is present
The default behavior of eslint --config is to use that config _in
addition to_ any default config (e.g. .eslintrc), not _instead of_. To
prevent this, you have to additionally pass --no-eslintrc.

We find this behavior surprising and prefer setting a custom config to
also imply --no-eslintrc, despite being a behavior difference between
the engine and the underlying tool.