Fix `make test` and `make integration`
We're hitting errors where the `mocha` command is not available. I
believe this is because we were running `yarn` in `/usr/src/app`, and
then trying to run tests in `/code`, where the node_modules aren't
available.

For a sec, I assumed this was related to
https://github.com/codeclimate/codeclimate-eslint/pull/256 but I
confirmed the issue predates that change.